refactor(transport): decompose client.py into client/ package (S2)
Extract _parse_retry_after + _log_circuit_event into _helpers.py; send shrinks to 157-LOC orchestration. Deduplicate get_manifest/discover via shared _fetch_and_cache_manifest core. Split ASAPClient into _core/_send/_discovery/_helpers; __init__ re-exports ASAPClient + re-binds asyncio/time/httpx/logger so existing patch() seams keep resolving.
